Use parallel MySQL database, reasoning in multiline comment
Turned out PostgreSQL doesn't support absolute table locking, and since we want: 1) keep using Django ORM and 2) an absolute table lock during ratelimit checking; We need something else for this than PostgreSQL. MySQL supports this by LOCK TABLE ... WRITE, so setting up a MySQL specificly for this purpose. If it'll be refactored to something else later, maybe, but for now it seems the most optimal solution.